+++ src/components/Swap.js
@@ -7,12 +7,31 @@ import Button from 'react-bootstrap/Button';
 import Row from 'react-bootstrap/Row';
 import { ethers } from 'ethers';
 
+import AMMabi from '../abis/AMM.json';
+
 const Swap = () => {
+  const [isApproved, setIsApproved] = useState(false);
   const [inputAmount, setInputAmount] = useState('');
   const [outputAmount, setOutputAmount] = useState(null);
   const { account } = useContext(BlockchainContext);
   const provider = new ethers.providers.Web3Provider(window.ethereum);
   const signer = provider.getSigner(account);
+
+  const [ammContract, setAmmContract] = useState(null);
+
+  useEffect(() => {
+    const initializeContract = async () => {
+      if (typeof window.ethereum !== 'undefined' && account) {
+        const provider = new ethers.providers.Web3Provider(window.ethereum);
+        const signer = provider.getSigner();
+        const contractAddress = '0x5d4bf065A9ae8B5D8b9cb632C38c77d784d086Cc';
+        const contract = new ethers.Contract(contractAddress, AMMabi, signer);
+        setAmmContract(contract);
+      }
+    };
+
+    initializeContract();
+  }, [account]);
 
   const [currency, setCurrency] = useState({ input: 'USDC', output: 'EURC' });
 
@@ -69,6 +88,21 @@ const Swap = () => {
     // Implement swapping logic
   };
 
+  const handleApprove = async () => {
+    if (ammContract && account) {
+      try {
+        await ammContract.setSwapApproval("0xFC116bF2F327b8A0DFba8fB49a9Eb8BC041D165E", true);
+        setIsApproved(true);
+      } catch (error) {
+        console.error("Error approving swap:", error);
+      }
+    } else {
+      console.error("AmmContract or account not available");
+    }
+  };
+
+  
+
   return (
     <Card style={{ maxWidth: '450px' }} className="mx-auto px-4">
       <Form onSubmit={swapHandler} style={{ maxWidth: '450px', margin: '50px auto' }}>
@@ -103,6 +137,24 @@ const Swap = () => {
             Switch Currency
           </Button>
         </Row>
+
+        <Card style={{ maxWidth: '450px' }} className='mx-auto px-4'>
+        <Card.Body>
+          <Card.Title>Approve Swap</Card.Title>
+          <Card.Text>
+            {isApproved 
+              ? "Swap has been approved for the contract."
+              : "Click the button below first to approve the contract for swapping."}
+          </Card.Text>
+          <Button 
+            variant="dark" 
+            onClick={handleApprove}
+            disabled={isApproved || !account || !ammContract}
+          >
+            {isApproved ? "Approved" : "Approve Contract for Swap"}
+          </Button>
+        </Card.Body>
+      </Card>
 
         <Row className="my-3">
           <Button type="submit" variant="dark">Swap</Button>
